+++ sdk/storage/storage-blob/src/Clients.ts
@@ -79,7 +79,8 @@ import {
   ModificationConditions,
   ModifiedAccessConditions,
   BlobQueryArrowField,
-  BlobImmutabilityPolicy
+  BlobImmutabilityPolicy,
+  HttpAuthorization
 } from "./models";
 import {
   PageBlobGetPageRangesDiffResponse,
@@ -113,6 +114,7 @@ import {
   extractConnectionStringParts,
   generateBlockID,
   getURLParameter,
+  httpAuthorizationToString,
   isIpEndpointStyle,
   parseObjectReplicationRecord,
   setURLParameter,
@@ -616,6 +618,10 @@ export interface BlobSyncCopyFromURLOptions extends CommonOptions {
    * Blob tags.
    */
   tags?: Tags;
+  /**
+   * Only Bearer type is supported. Credentials should be a valid OAuth access token to copy source.
+   */
+  sourceAuthorization?: HttpAuthorization;
 }
 
 /**
@@ -1779,6 +1785,7 @@ export class BlobClient extends StorageClient {
           sourceIfUnmodifiedSince: options.sourceConditions.ifUnmodifiedSince
         },
         sourceContentMD5: options.sourceContentMD5,
+        copySourceAuthorization: httpAuthorizationToString(options.sourceAuthorization),
         blobTagsString: toBlobTagsString(options.tags),
         immutabilityPolicyExpiry: options.immutabilityPolicy?.expiriesOn,
         immutabilityPolicyMode: options.immutabilityPolicy?.policyMode,
@@ -2475,6 +2482,10 @@ export interface AppendBlobAppendBlockFromURLOptions extends CommonOptions {
    * Storage Services.
    */
   encryptionScope?: string;
+  /**
+   * Only Bearer type is supported. Credentials should be a valid OAuth access token to copy source.
+   */
+  sourceAuthorization?: HttpAuthorization;
 }
 
 /**
@@ -2887,6 +2898,7 @@ export class AppendBlobClient extends BlobClient {
           sourceIfNoneMatch: options.sourceConditions.ifNoneMatch,
           sourceIfUnmodifiedSince: options.sourceConditions.ifUnmodifiedSince
         },
+        copySourceAuthorization: httpAuthorizationToString(options.sourceAuthorization),
         cpkInfo: options.customerProvidedKey,
         encryptionScope: options.encryptionScope,
         ...convertTracingToRequestOptionsBase(updatedOptions)
@@ -3033,6 +3045,10 @@ export interface BlockBlobSyncUploadFromURLOptions extends CommonOptions {
    * Optional. Conditions to meet for the source Azure Blob.
    */
   sourceConditions?: ModifiedAccessConditions;
+  /**
+   * Only Bearer type is supported. Credentials should be a valid OAuth access token to copy source.
+   */
+  sourceAuthorization?: HttpAuthorization;
 }
 
 /**
@@ -3256,6 +3272,10 @@ export interface BlockBlobStageBlockFromURLOptions extends CommonOptions {
    * Storage Services.
    */
   encryptionScope?: string;
+  /**
+   * Only Bearer type is supported. Credentials should be a valid OAuth access token to copy source.
+   */
+  sourceAuthorization?: HttpAuthorization;
 }
 
 /**
@@ -3831,6 +3851,7 @@ export class BlockBlobClient extends BlobClient {
           sourceIfTags: options.sourceConditions?.tagConditions
         },
         cpkInfo: options.customerProvidedKey,
+        copySourceAuthorization: httpAuthorizationToString(options.sourceAuthorization),
         tier: toAccessTier(options.tier),
         blobTagsString: toBlobTagsString(options.tags),
         ...convertTracingToRequestOptionsBase(updatedOptions)
@@ -3928,6 +3949,7 @@ export class BlockBlobClient extends BlobClient {
         sourceRange: offset === 0 && !count ? undefined : rangeToString({ offset, count }),
         cpkInfo: options.customerProvidedKey,
         encryptionScope: options.encryptionScope,
+        copySourceAuthorization: httpAuthorizationToString(options.sourceAuthorization),
         ...convertTracingToRequestOptionsBase(updatedOptions)
       });
     } catch (e) {
@@ -4703,6 +4725,10 @@ export interface PageBlobUploadPagesFromURLOptions extends CommonOptions {
    * Storage Services.
    */
   encryptionScope?: string;
+  /**
+   * Only Bearer type is supported. Credentials should be a valid OAuth access token to copy source.
+   */
+  sourceAuthorization?: HttpAuthorization;
 }
 
 /**
@@ -5070,6 +5096,7 @@ export class PageBlobClient extends BlobClient {
           },
           cpkInfo: options.customerProvidedKey,
           encryptionScope: options.encryptionScope,
+          copySourceAuthorization: httpAuthorizationToString(options.sourceAuthorization),
           ...convertTracingToRequestOptionsBase(updatedOptions)
         }
       );