+++ sdk/storage/storage-blob/src/index.ts
@@ -34,7 +34,8 @@ export {
   ObjectReplicationRule,
   ObjectReplicationStatus,
   BlobQueryArrowField,
-  BlobQueryArrowFieldType
+  BlobQueryArrowFieldType,
+  HttpAuthorization
 } from "./models";
 export * from "./Pipeline";
 export * from "./policies/AnonymousCredentialPolicy";