+++ sdk/storage/storage-blob/test/node/blobclient.spec.ts
@@ -8,7 +8,7 @@ import { join } from "path";
 
 import { AbortController } from "@azure/abort-controller";
 import { isNode, TokenCredential } from "@azure/core-http";
-import { delay, record, Recorder } from "@azure/test-utils-recorder";
+import { delay, isPlaybackMode, record, Recorder } from "@azure/test-utils-recorder";
 
 import {
   BlobClient,
@@ -26,6 +26,8 @@ import {
   getBSU,
   getConnectionStringFromEnvironment,
   getImmutableContainerName,
+  getTokenBSU,
+  getTokenCredential,
   recorderEnvSetup
 } from "../utils";
 import { assertClientUsesTokenCredential } from "../utils/assert";
@@ -203,6 +205,91 @@ describe("BlobClient Node.js only", () => {
       result3.segment.blobItems![1].properties
     );
     assert.ok(result3.segment.blobItems![0].snapshot || result3.segment.blobItems![1].snapshot);
+  });
+
+  it("syncCopyFromURL - source SAS and destination bearer token", async function() {
+    if (!isPlaybackMode()) {
+      // Enable this when STG78 - version 2020-10-02 is enabled on production.
+      this.skip();
+    }
+    const newBlobName = recorder.getUniqueName("copiedblob");
+    const tokenBlobServiceClient = getTokenBSU();
+    const tokenNewBlobClient = tokenBlobServiceClient
+      .getContainerClient(containerName)
+      .getAppendBlobClient(newBlobName);
+    const newBlobClient = containerClient.getBlobClient(newBlobName);
+
+    // Different from startCopyFromURL, syncCopyFromURL requires sourceURL includes a valid SAS
+    const expiryTime = recorder.newDate("expiry");
+    expiryTime.setDate(expiryTime.getDate() + 1);
+
+    const factories = (containerClient as any).pipeline.factories;
+    const credential = factories[factories.length - 1] as StorageSharedKeyCredential;
+
+    const sas = generateBlobSASQueryParameters(
+      {
+        expiresOn: expiryTime,
+        permissions: BlobSASPermissions.parse("racwd"),
+        containerName,
+        blobName
+      },
+      credential
+    );
+
+    const copyURL = blobClient.url + "?" + sas;
+    const result = await tokenNewBlobClient.syncCopyFromURL(copyURL);
+    assert.ok(result.copyId);
+
+    const properties1 = await blobClient.getProperties();
+    const properties2 = await newBlobClient.getProperties();
+    assert.deepStrictEqual(properties1.contentMD5, properties2.contentMD5);
+    assert.deepStrictEqual(properties2.copyId, result.copyId);
+  });
+
+  it("syncCopyFromURL - destination bearer token", async function() {
+    if (!isPlaybackMode()) {
+      // Enable this when STG78 - version 2020-10-02 is enabled on production.
+      this.skip();
+    }
+    const newBlobName = recorder.getUniqueName("copiedblob");
+    const tokenBlobServiceClient = getTokenBSU();
+    const tokenNewBlobClient = tokenBlobServiceClient
+      .getContainerClient(containerName)
+      .getAppendBlobClient(newBlobName);
+    const newBlobClient = containerClient.getBlobClient(newBlobName);
+
+    const result = await tokenNewBlobClient.syncCopyFromURL(blobClient.url);
+    assert.ok(result.copyId);
+
+    const properties1 = await blobClient.getProperties();
+    const properties2 = await newBlobClient.getProperties();
+    assert.deepStrictEqual(properties1.contentMD5, properties2.contentMD5);
+    assert.deepStrictEqual(properties2.copyId, result.copyId);
+  });
+
+  it("syncCopyFromURL - source bearer token and destination account key", async function() {
+    if (!isPlaybackMode()) {
+      // Enable this when STG78 - version 2020-10-02 is enabled on production.
+      this.skip();
+    }
+    const newBlobName = recorder.getUniqueName("copiedblob");
+    const newBlobClient = containerClient.getBlobClient(newBlobName);
+
+    const tokenCredential = getTokenCredential();
+    const accessToken = await tokenCredential.getToken([]);
+
+    const result = await newBlobClient.syncCopyFromURL(blobClient.url, {
+      sourceAuthorization: {
+        scheme: "Bearer",
+        parameter: accessToken!.token
+      }
+    });
+    assert.ok(result.copyId);
+
+    const properties1 = await blobClient.getProperties();
+    const properties2 = await newBlobClient.getProperties();
+    assert.deepStrictEqual(properties1.contentMD5, properties2.contentMD5);
+    assert.deepStrictEqual(properties2.copyId, result.copyId);
   });
 
   it("syncCopyFromURL", async () => {